backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
python-dotenv
requests
redis
python-multipart

openai
PyJWT
celery[redis]
python-dotenv
